move Sample enum to pipeline folder
