+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/control/vision/pipeline/Sample.java
@@ -1,0 +1,7 @@
+package org.firstinspires.ftc.teamcode.control.vision.pipeline;
+
+public enum Sample {
+    NEUTRAL,
+    BLUE,
+    RED,
+}
