--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystem/Sample.java
@@ -1,7 +1,0 @@
-package org.firstinspires.ftc.teamcode.subsystem;
-
-public enum Sample {
-    NEUTRAL,
-    BLUE,
-    RED,
-}
